A quick lesson on why you should forget about using f32
1713474759893 converted to f32 is 1713474764800. The machine error resulting from this conversion is 4907.

+++ turbofuro_runtime/src/actions/time.rs
@@ -50,7 +50,7 @@ pub async fn get_current_time<'a>(
     step_id: &str,
     store_as: Option<&str>,
 ) -> Result<(), ExecutionError> {
-    let value: StorageValue = (Utc::now().timestamp_millis() as f32).into();
+    let value: StorageValue = (Utc::now().timestamp_millis() as f64).into();
     store_value(store_as, context, step_id, value).await?;
     Ok(())
 }
